About & Recipes pages
Hi ladies, here are the files I have modified. I also added the ext_api.py file because I had to modify line 24. I basically sliced the image url  string returned so that we can get the exact url required. May have to do it for others such as saved-recipes, but will deal with that when we get there.